Fix insiduous caching bug
We saw the behaviour that after changing block contents and saving those changes were not visible.

The reason was that the @@contents view came with a Cache-Control of 60 seconds -- which definitely should not be there.

The @@contents view uses the @@object-details view internally (via the pagetemplate), which in turn sets cache-control (which makes some sense when @@object-details is requested directly, since it mostly returns images or other referenced content) -- but the cache-control header got set anyway, even when the view was called by the pagetemplate. Hell in a handbasket.

We decided that since dnd.js implements its own (JS-based) cache for object details, the cache-control header isn't really necessary or useful, anyway, so we removed it.