Introduce IFoldable marker interface for foldable blocks
We don't really know why the foldable stuff did not show up before (while the viewlet was registered in zeit.content.article for IElement), but once I started moving the folding stuff to zeit.edit, it started showing up everywhere (as it should have, from the beginning, since it was registered for IElement). Strange.

Using an explicit interface clears up this confusion.