re #8996: changed object names of videos and playlists to ease refactoring of brightcove repository
We'd like to break the refactoring of the brightcove repository down into
small and understandable steps. One step is to change the storage from BTree
to DAV. While doing that, we will need to address objects by a xml.zeit.de
unique ID which, being a URL, doesn't allow for a ':' within the object's
name.