Mark transactions undoable only if the form may actually be writing something
Marking in call leads to unexpected side effects, since the first time a checked-out article is displayed, its blocks are assigned their UUIDs (so an transaction that /should/ be read-only actually writes things).

Thus we mark in the success handler, because we can be reasonably sure that there will be a real, undoable commit coming up at that point in time.

Note that if the transaction doesn't write anything (because nothing has changed), marking the transaction has no effect. And since formlib makes it neigh impossible to determine that, we don't go to that trouble.