don't submit while mouse is down
On a widget with a calendar or the +1W button, clicking those buttons sends the
focusout event. This yields to the form being saved *before* the (+1W) data can
be set to the field. After the form is reloaded the data is gone. Thus the
buttons don't work. Preventing save during mouse down fixes this.

fixes #12119