load editor when no specific id is given
This interacts with makeJSONRequest, which reloads the `target_component` when
no specific signals are sent from the server. LoadAndReload in turn doesn't
pass a `target_component`, which should default to the whole editor being
reloaded. makeJSONRequest falls indeed back to reloading the whole editor when
no `target_component` is given but the editor didn't understand the signal
anymore.