Re #11421: Restructure tagging widget not to use CheckboxWidget as base
The paradigm of CheckboxWidget is to display all items of a set and choose some of them,
this does not match the new requirement of adding additional tags (from the whitelist).
Additionally, CheckboxWidget would not lend itself well to transporting further data,
such as pinned tags (#12521).

Thus we discard the CheckboxWidget base and just write our own widget class.