Re #12573: Provide WEBDRIVER_LAYER side-by-side to SELENIUM_LAYER
We want to use the same SeleniumTestCase regardless of layer, so we
need to take care that everything we do there is written in way that
is compatible with both.

Thus we use WebdriverSeleneseTestCase as the base class, which is the
same as RCTestCase, except that it does not call
self.selenium.setContext in setUp, which does not exist in Webdriver
anymore.