Re VIV-252: Don't use BeforeAddedEvent to serialize provided interfaces to DAV
This proves to be too decoupled, since we will need to add the section
marker interfaces in that precise event -- which wouldn't work because
the order of event handlers is arbitrary.

But since storing interfaces can be considered quite the core behaviour,
it's alright to move this in closer to the serialization process instead of
using events and pretending it's decoupled.