Re VIV-276: Change XMLReference for image to use XMLReferenceUpdater
Previously the code relied on the assumption that the XMLReference
nodes are recreated each time, which does not hold anymore since we
leave the nodes alone so as not to disturb properties of reference
objects.

Thus, we use the same pattern here as with RelatedContent (which makes
more sense anyway), where the actual updating of metadata is done by
an XMLReferenceUpdater, and creating XMLReference simply delegates to
that.