Re VIV-282: Allow referencing the same object several times for MultiResource
Since Reference objects only have (source, attribute, target) as their
uniqueId, they don't support referencing the same object multiple
times. Usually, this is a sensible restriction, but at least the feed
of centerpages also uses MultiResource, and there it's necessary and
sensible to reference the same article multiple times (in different
teaser block, for example).

Or, put differently, we do currently not have the use case of wanting
to access the same ReferenceProperty as a MultiResource, too. So we
lifted the "same object only once" restriction for MultiResource.