Re VIV-305: Determine the correct XML node for SingleReferenceProperty
The lxml.objectify API offer an insiduous source of bugs: Iterating
over a single element a) is possible and b) yields all siblings with
the same tag. So it has been easy for SingleReferenceProperty to
overlook the fact that when we find a single element via xpath, that
already is the one we want.

A test for this is in zeit.content.article, since it needs an xpath of '.'
and that is rather difficult to set up in isolation