Re VIV-322: Update metadata of free teasers correctly on checkin
Previously, free teasers contained the teaser-id in @href, so even though
cms_content_iter yields both the teaser and the referenced object, only
the teaser was found by updateMetadata.

Now, @href always is the referenced uniqueId, so updateMetadata needs
to handle free teasers explicitly. This is slightly bad from a coupling
perspective, but since updateMetadata bypasses objects and directly works
on the XML, that is to be expected (since the encapsulation is already broken
at that point).