Fixes VIV-377: Don't store empty string values for IImageMetadata.caption
This bug was an accidental collision of changes.
First, zeit.cms commit:aa79d6dc0a8b introduced missing_value='' to fix problems with XMLSnippet
Later, zeit.cms commit:d89d3c846768 removed XMLSnippet completely, but left the missing_value in

which leads to the incorrect behaviour of storing empty strings with the new field type Text,
which then in turn causes empty image captions to be written when a user does not fill
in the caption (but only changes the ``alt`` value, for example).