Re VIV-405: Add event when dragging modules, so the content-editable can save itself and restore its landing zones
Note that the content-drag start handler in zeit.cms.browser:js/dnd.js
calls Droppables.prepare() synchronously instead of in an
after-reload handler, which is the proper way we use here.
It seems that the timing is more favorable in the content case than in
the module case, so that it works out accidentally -- kludgy! And
since dnd.js is in zeit.cms, we can't actually depend on
zeit.edit.editor, so we're *lucky* it works, otherwise we'd have to do
something clever and/or ugly to resolve the dependency.