Re DEV-797: Create thumbnail source image only while group is in repository
Apache's mod_dav doesn't seem to allow adding a member to a locked
collection, unfortunately, even though the WebDAV spec reads like that
should be possible. Specifically: calling LOCK on a member inside a
collection checks the parent collection with the flag "acquire
exclusive lock", which is of course denied since the collection
already is locked; however my understanding is that the client wants
the lock only for the member, not the parent, so the check about the
state of the parent should reflect that.