Re DEV-797: Move the variant UI to repository group
We cannot create a thumbnail source while the group is locked (see
commit:f38df14 for details), but we want to reflect a changed master
image immediately. This way, the variant UI (that creates the
thumbnail source) is only accessible while checked in, so no lock.

(An alternative would be to store the thumbnail source in the ZODB
instead of the repository, however that would benefit only the variant
UI, and not e.g. the CP editor, which also needs to generate a lot of
variant images and thus needs the performance boost, too)