Add parameter ``use_default_value`` to ``ObjectPathAttributeProperty``, and generally make it behave more like ``ObjectPathProperty``.
This is a behaviour change: previously it would return the
missing_value if the XML contained invalid data (e.g 'foo' if the
field is an Int); now this will raise ValueError.

I'm not sure why we ever would have the two properties differ in their
behaviour this much; we'll see if something depends on this strange behaviour.