Bugfix: actually create thumbnail source when an imagegroup is added.
We only add the thumbnail source when the imagegroup is unlocked --
but only IRepositoryContent has lock information at all. The fact
that we *do have* repository content was lost by the imagroup AddForm,
since it called add() some more to add images, and then overwrote
`_created_object` with the not-yet-wrapped imagegroup object.