FIX: Send no-cache headers at least for our full pages, so browsers don't cache e.g. a directory listing
I don't have the ambition at the moment that this should definitely
apply absolutely everywhere. I've checked that it will apply to
basically all full-page views and forms (our view class structure is
quite simple, so this and the zeit.edit main view is actually the only
change needed), which is where the main pain point was discovered.