TMS-170: Explicitly index TMS keywords
We no longer have the use case that we want to transmit Intrafind
keywords to TMS; during bulk indexing we update the keywords from TMS,
keeping only the pinned ones.

And since we want to index to TMS in parallel in production while
Intrafind is still active, we have to make sure to use the TMS
keywords there too (otherwise changes via vivi-beta have no effect).