TMS-200: Index TMS during publishing too, so it immediately receives the "published" timestamp information for newly published objects
This was a performance optimization to speed up publishing which we
inherited from zeit.solr, but is not actually applicable for us here.
The ext. solr is updated by zeit.publisher, so it's fine there, but
here TMS is updated by vivi, so publishing needs to be considered as well.