TMS-227: Remove attempts for UI combining different fields into one query field
Since we'll want to display different widgets (e.g. dropdown for
channels, or series, or freetext entry for keywords), putting this
into one combination field is not feasible. We'll probably break out
into different automatic_type values instead (but maybe we'll
serialize all of them to `<query type="foo"`> nodes).

This also clarifies that we won't attempt to port the fledgling
"keyword" idea to ES in TMS-219.