TMS-231: Cache area `values()` so they don't consume extra TMS results.
`filter_values` for example iterates the area's `values()` again when
called via `existing_teasers`. This has simply cause additional queries
before, but with the centrally cached result sets, this now consumes
results, effectively changing the order in which they are assigned to
the page's (automatic) areas.

Please note that this also introduces a central cache on the centerpage,
which can then be used to cache the various things we need (on the
transaction boundary).