ZON-4106: Add js to dynamically render widgets
The select box value determines which input fields are rendered below.

In read only mode we will still render all possible fields, since the
DOM varies a lot from the write mode. While it is technically possible
to hide fields in ro mode as well, it requires lots of ugly DOM
navigation to accomplish this, so it is omitted as of now.