ZON-4106: Revert widget order
This reverts most of 24e51d5cb59185ae7c07ba63666759a309c24cfe.

Unfortunately we cannot reorder fields in the model, because it will
mess with all existing content, not being mapped properly to the new
field order. Instead, all existing data would just be put over as is, in
the stored (read: previous) order over the new field order, which
results in fields having not their belonging data.

Moreover in some cases the code relies on a specific order, which we
would need to rewrite to work with both order models. This sounds like a
lot of headache!
See: metadata.ImageMetadata.copyrights