ZON-4106: Manipulate DOM through js instead of relying on css alone
Unfortunately moving tablerows around with relative positioning is not
precisely predictable across different firefox versions on different
systems. Since even a single pixel difference in the checked-out state
will be noticeable, as the widgets photographer and copyrights freetext
overlap, we instead have to manipulate the DOM with javascript!

Since such a manipulation is even more messy in the checked-in state and
we do not rely on widget placement by pixel in this case, it is fine to
use relative positioning for the checked-in state.