ZC-84: Use helper for mocking the account server, thereby simplifying the tests.
Note that this also changes the tests subtly. They — except for the
new one — now test the changes on the original content item, which
does not involve the update event subscriber. That is only invoked
while editing the checkout out version and to test this the events
fired by `zope.formlib` need to be "simulated".