MAINT: Guess MIME type in filesystem connector just like Apache does for DAV connector
I'm not sure what the original idea for mock connector had been, to
cache the mime type on writes. Since there is no heuristics involved
on write anywhere, this mostly yielded `application/octet-stream`
(which we had as an assertion in a few tests in zeit.cms, where it
semantically made no sense at all).

In production with DAV, the `contenttype` property is filled in by
Apache (using its own libmagic module).