PERF: Determine image mime type only on demand, not always upfront on resolve
This saves a lot of unnecessary effort, e.g. when we only resolve an
object to read its workflow properties or some such, and are not
interested in its intrinsic properties at all. Also, while we only
have read the first few bytes for the identification, zeit.connector
always has to read the whole body, which then uses quite a bit of RAM.

Since be8134d we don't set the mime type explicitly anyway, and only
use detection, so we don't lose any functionality here. We do lose the
safetybelt of "don't even resolve it if it's not an image/ mime type",
but the only situation where that could be relevant is data
corruption, and then an exception later on is fine as well, I think.