FIX: Unregister patched utilities on test teardown
zcapatch stores the registry that was active when it was constructed, which is
not actually all that helpful for us, since we construct it quite early in the
layer. Not sure why this has never been an issue before; if it pops up more
ofthen we'll have to solve this generically somehow.