OPS-1133: Implement a DAV cache (properties and childnames) with dogpile/redis as the storage backend
This will not be used by default, since the utilities for the
zeit.connector caches are currently persisted in the ZODB. So you have
to manually change the utility registrations to activate it. (If/when
we're satisfied that this is the way to go, we'll change this into a
non-persistent registration instead).