OPS-1133: Modify DAV cache conflict resolution rules to avoid deleting cache entries
Deleting was definitely correct, but it caused thundering herd issues
e.g. for often-used folders. We now switch to a simple "last request
wins" for properties (which hopefully does not wrongly overwrite
previous changed -- since concurrent accesses to the same object
should be locked anyway). For childnames we go back to
ConflictError/retry (since deleting would cause the folder properties
to be deleted as well, to make the cache consistent), and hope that
folder childnames to not conflict often.