WOMA-70: Provide endpoint for ingredient search
This is a revival of code, that has once been part of zeit.intrafind,
with some specific adaptations for ingredients and modernizations to be
python 3 ready.

An endpoint is provided as @@ingredients_find, which will search through
a XML source. Searching for word stems is possible, to make autocomplete
available as well.

See: https://github.com/ZeitOnline/zeit.intrafind/blob/master/src/zeit/intrafind/whitelist.py