STO-49: Support setting a default filename for source config files
We no longer require an explicit path/url to be configured, but rather
default to a central `config-base-url` setting if no explicit value is
configured. This should vastly reduce the amount of boilerplate
configuration needed (e.g. to differentiate between staging and
production), since for most cases only the base-url has to be
overridden now.