ZON-6068: Fix template->header->color hierarchy TTW
this has only been tested so far locally, proper tests to follow

note: the masterslave widget makes the implicit assumption that its
``attribute`` attribute is valid both for the slave AND the widget.

We either need to expand the code to remove this assumption or ensure in
our markup, that it's the same throughout (i.e. ``name`` in this case)