STO-59: Upgrade the `elasticsearch` Python client library
The library should actually match the Elasticsearch (major) version
running on the server, but we want to talk to version 2.x and 7.x at
the same time and we're only using the `search` API method. So for
the time of the migration to 7.x we assume it should work with both.