STO-185: Also cache DAV properties on file modification time
Since the cache is only active for filesystem connectors, in which case
all DAV properties are also stored in the filesystem, the property
values can only change if the underlying file itself is changed as well.

Note that this assumes the `.meta` files is always updated along with
the file representing the content item itself and vice versa.