STO-185: Refactor path and mtime handling (WIP)
The idea is that the cache itself doesn't need to know about any of
the connector's internals like the repository path etc. This is not
finished, though, and a number of tests are still failing. A good
starting point to debug this is:

  $ bin/test vivi/core/src/zeit/cms/repository/

Those are the only failures within `zeit/cms/`