ZON-6539: Adjust test content for DailyNL prop
In addition to: 0ea71e536541e9cbaf32b5fc8469474a2150c9d7

We need to keep the DailyNL prop for one conversion test, since only the
form field is removed, but the dav property is still available in
previous articles. Thus we need to ensure conversion still works for
those as well.