TOPIC-31: Provide kpi for reach content
This is a little fuzzy from a modelling perspective, unfortunately,
since reach results only ever contain a single score, for the metric
that was requested; while IKPI implies that it provides any and all kpis.
Since that's not possible with reach results, we fold this into a
single `score` attribute, to at least make clear what is happening.