ZO-788: Make Passthrough connector work with migration
The IResource API is a bit dodgy, sometimes .data is a fresh stream on
each access, sometimes you'd have to seek(0) first, and so on. So
let's sidestep that issue by not reusing the resource object at all,
after importing it, and instead fetch a fresh one (which comes from
transaction cache anyway, so it should be cheap enough).