MAINT: Publish breaking news banner directly together with its article
Previously this was quite the hack, shoehorning a second transaction
into the tail-end of the currently running one (to update the DAV
cache with banner/enabled=False, but only if the main transaction was
successful).

transaction-3.x finally plugs the loophole we've been exploiting to do
this, so we combine both actions (publish article and publish banner)
into the same transaction to achieve the desired outcome.

A little problematic is that we cannot write a proper test here, since
the mock connector behaves differently regarding transactions than the
DAV+zodb cache one, sigh.