ZO-1858: Speed up bw-compat code for image group without master images
The code from a2954b2 wanted to get the largest image, but always had
a bug and just took the first one. Since this does not seem to have
mattered at all, we make that behaviour explicit, which now yields a
significant performance benefit, as we do not have to get the image
body (separately from GCS) to determine the size -- once we store that
information in properties, we should revert this (and implement it
properly of course).