ZO-2874: Opening and closing characters can't be identified
Since all opening an closing quote characters are used differently in
text editors, we need to come up with a rule, that identifies this more
accurate. For this we assume, that an opening quote sign is always
followed by a alphanumerical sign and a closing quote is always preceded
by a alphanumerical sign.