MAINT: Simplify spelling
This is rather old, but IIRC the issue was that `reload()` did not
work and an explicit `open()` was required. Meanwhile,
zope.testbrowser has long switched to webtest from mechanize, and now
it works just fine. (Even though it occurs to me that zope.formlib
does not perform "redirect after POST", so maybe this runs POST twice
now? Doesn't much matter for the tests, though.)