ZO-638: Provide some UI feedback about push sending
We're quite aware of the limitations of this approach, but it would be
quite finicky to have vivi retrieve errors from the publisher job
queue after the fact. So the current perspective is that sending push
notifications should be its own action, and not bundled together with
publishing anymore, which would solve all problems we're trying to
adress in this issue (transaction boundaries, retry prevention, UI
feedback etc).