WCM-271: Pass property ns+name to IDAVPropertyConverter
We need to differentiate properties that already have been migrated to
columns from those that still remain in unsorted, for the transition
period. And we can only determine this information via ns+name,
especially in the `toProperty` case (in `fromProperty` we might have
annotated the value returned by connector, or something).