ZO-5176: Use correct primary key for paths table: id, not parent+name
This has always been incorrect (see also description of ZO-5176),
but did not have many concrete consequences (the only real one was sqlalchemy
complaining about bulk updates in move()) so far.

But since we now want to lean on the sqlalchemy session identity map
to pre-populate the content/path relationship, we need the correct key
(content uuid) there, otherwise the Path objects cannot be looked up.

Since we're going to delete the paths table entirely in the next step,
we're intentionally not updating the primary key definition in existing databases.
I'm quite certain that it does not really matter anyway:
* sqlalchemy does not change its behaviour much (except for the identity map, which we *want*)
* we never want to have 2 paths entries for 1 content (which would be prevented by id being the primary key)
* the parent/path/id combination unique constraint on paths is wrong anyway (should only be parent+path)