WCM-20: Combine fields with AND/OR
Note that sqlalchemy seems to generate as few parens as possible, so e.g.
there are some for `(channels @> '...') OR (channels @> '...')`, but none in
`ressort = 'International' AND sub_ressort = 'Nahost' OR ressort = 'Wissen'`.
I do think this is _correct_, because AND has precendence over OR in SQL,
but it may be a bit surprising to read.