WCM-271: Simplify mock connector type conversion
As a85dbd0 shows, the original idea in 14869f5 (let's make an easy
code structure, always deserialize on read, always serialize on write)
was mistaken, because it did not take into account non-connector
clients, i.e. writes by the content layer, which store typed
properties when the `write` toggle is enabled.

Thus, if we already have to be careful about deserializing, there's
not point in additionally serializing anyway.