WCM-1210: Mark image non-required in scrolly modules, to avoid validation false positives
Yes, this is semantically incorrect. Yes, other modules (e.g. image) use the wrong declaration too.

However, due to having inherited the validation mechanics from z.c.cp,
they only are evaluated when the whole module is reloaded, which in
z.c.article does not happen regularly, because the inline form only
reloads "itself", but not the whole module.

So in the common case, where the user drop a published image and
everything is fine, this way it's less confusing to not show the
yellow mark at all (even though this may "hide" error cases of
forgetting to drop an image or using an unpublished one).